requirements.txt
websocket-client>=1.8.0
pynput>=1.7.6
